`Added MediatR and Employee entity with related classes and migration`
Implemented the MediatR package in the `CQRS_Project.csproj` file to support the mediator pattern. Created a new `Employee` entity in the `Context.cs` and `ContextModelSnapshot.cs` files, and added a corresponding `Employee` class. Also, introduced `GetEmployeeQueryHandler`, `GetEmployeeQuery`, and `GetEmployeeQueryResult` classes to handle queries for employee data. Lastly, a new migration `Mig2` was added to create an `Employees` table in the database, and the `Employee` entity was included in the model builder in the `20240606110610_Mig2.Designer.cs` file.